refactored temporary_patches.py to temporary_patches folder and added csm merge inputs patch to fix eos token training
